Updated and added script to Blood Port
Made a batch script that when pointed to the directory of the games it'll add the games needed under the name "AddFilesTodirectory.bat" Port also needed to put the needed for the Blood.INI file to work. Added it to the list of needed files. Due to :
SOUNDS.RFF & blood.rff are .RFF files. *RFF already covers the need for those files. So removed from the list.

Now works well on the rgb30.